🔨 refactor : auth 폴더 global로 이동
