--- src/main/java/com/example/cloudfour/peopleofdelivery/domain/auth/dto/LoginRequestDto.java
@@ -1,4 +1,0 @@
-package com.example.cloudfour.peopleofdelivery.domain.auth.dto;
-
-public class LoginRequestDto {
-}
