+++ src/main/java/com/example/cloudfour/peopleofdelivery/global/auth/dto/OAuthRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.cloudfour.peopleofdelivery.domain.auth.dto;
+package com.example.cloudfour.peopleofdelivery.global.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
