+++ src/main/java/com/example/cloudfour/peopleofdelivery/global/auth/dto/OAuthResponseDTO.java
@@ -1,0 +1,4 @@
+package com.example.cloudfour.peopleofdelivery.global.auth.dto;
+
+public class OAuthResponseDTO {
+}
