+++ src/main/java/com/example/cloudfour/peopleofdelivery/global/auth/oauth/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.example.cloudfour.peopleofdelivery.domain.auth.oauth;
+package com.example.cloudfour.peopleofdelivery.global.auth.oauth;
 
 import com.example.cloudfour.peopleofdelivery.domain.user.entity.User;
 import org.springframework.security.core.GrantedAuthority;
